Tweak handling of parser errors
Once I got to actually using synchronize(), I realized I had kinda
totally missed the point of all the errors in the parser: we want to
report as many as we can, not just the first one we see! Instead of
using a global like the book does, we'll accumulate errors on the
parser, and print them all out before we try running anything.

I'm not totally excited about this because the caller now needs to know
to check the parser, but I'll tweak that here in a second.