interpreter: eval refs, do not consume ASTs
This was the cause of the weird cloning in the previous commit. Instead
of having all these eval methods consume the AST nodes, they should get
references to them instead. This is much cleaner, and means that we can
evaluate an expression more than once without needing to .clone() it
every time, lol.