Implement function calls, plus the global clock()
Hey hey, look here! This was a decent amount of work, because we need to
wrap up Rust closures in some kind of LoxValue. I'm leaning on Arc here
to magically implement clone, so that I don't have to rearchitect the
whole thing in terms of references. I might wind up reorganizing some of
this code, but here I think it's fine.